requirements.txt
fastapi
uvicorn[standard]
openai-agents[litellm]
litellm
chromadb
pandas
matplotlib
python-dotenv

